move cmd/util to runtimeconfig/util
